rename runtime_common -> common
