+++ runtime/src/lib.rs
@@ -44,8 +44,8 @@ use sp_version::NativeVersion;
 use sp_version::RuntimeVersion;
 use xcm_config::XcmOriginToTransactDispatchOrigin;
 
-pub mod runtime_common;
-pub use runtime_common::{
+pub mod common;
+pub use common::{
 	AVERAGE_ON_INITIALIZE_RATIO, DAYS, HOURS, MAXIMUM_BLOCK_WEIGHT, MILLISECS_PER_BLOCK, MINUTES,
 	NORMAL_DISPATCH_RATIO, SLOT_DURATION,
 };
@@ -56,12 +56,12 @@ pub use sp_runtime::{MultiAddress, Perbill, Permill};
 // Polkadot imports
 use polkadot_runtime_common::{BlockHashCount, SlowAdjustingFeeUpdate};
 
-pub use runtime_common::{AccountId, Balance, BlockNumber, Hash, Signature};
+pub use common::{AccountId, Balance, BlockNumber, Hash, Signature};
 
 use weights::{BlockExecutionWeight, ExtrinsicBaseWeight, RocksDbWeight};
 
 // XCM Imports
-use runtime_common::Nonce;
+use common::Nonce;
 use xcm::latest::prelude::BodyId;
 
 /// The address format for describing accounts.
